documents/220224/requirements.txt
flask
google-cloud-firestore
gunicorn
requests
wtforms
flask_restful
google-cloud-pubsub